requirements.txt
asgiref
Django
djangorestframework
djangorestframework-simplejwt
pillow
PyJWT
sqlparse
